% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_canopy.R
\name{calculate_canopy}
\alias{calculate_canopy}
\title{calculate_canopy}
\usage{
calculate_canopy(data, canopy, nodata = NULL, dsm, visiblepoints)
}
\arguments{
\item{data}{Numeric. The input type of canopy raster.
data=1: percentage raster (that represents the percentage of
canopy area in each cell).
data=2: binary raster (that only uses two values to represent whether
canopy exists in each cell).}

\item{canopy}{Raster.}

\item{nodata}{Numeric. If the argument 'data' is 2, indicate the value of
cells that don't have any canopy. (it is defaulted as NULL)}

\item{dsm}{Raster. The digital surface model(DSM) that is used for
function 'calculate_viewshed' to calculate viewshed.}

\item{visiblepoints}{Dataframe. The viewshed calulated by
function 'calculate_viewshed'.}
}
\value{

}
\description{
calculate_canopy
}
