% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversity.R
\name{calculate_diversity}
\alias{calculate_diversity}
\title{calculate_diversity}
\usage{
calculate_diversity(land, viewshed, proportion = FALSE)
}
\arguments{
\item{land}{Raster. The raster of land use/land cover}

\item{viewshed}{Viewshed object \link{package "viewscape"}.}

\item{proportion}{logical, indicating if the percentage(\%) of each land use should
be returned. Default is FALSE.}
}
\value{
Dataframe. The output is the percentage(\%) of each type of land use
within a viewshed.
}
\description{

}
\examples{

}
\references{

}
