% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_feature.R
\name{calculate_feature}
\alias{calculate_feature}
\title{calculate_feature}
\usage{
calculate_feature(
  type,
  raster,
  dsm,
  r,
  viewpoint,
  offset_viewpoint,
  nodata = NULL
)
}
\arguments{
\item{type}{Numeric. The input type of canopy raster.
type=1: percentage raster (that represents the percentage of
canopy area in each cell).
type=2: binary raster (that only uses two values to represent whether
canopy exists in each cell).}

\item{raster}{Raster.}

\item{dsm}{Raster. The digital surface model(DSM) that is used to
compute viewshed.}

\item{nodata}{Numeric. If type = 2, it indicates the value of
cells that are empty. (it is defaulted as NULL)}
}
\value{
Numeric. The canopy area in the viewshed.
}
\description{
calculate_feature
}
