% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_landcover.R
\name{calculate_landcover}
\alias{calculate_landcover}
\title{calculate_landcover}
\usage{
calculate_landcover(landcover, dsm, visiblepoints, vegetation, imperviousness)
}
\arguments{
\item{landcover}{Raster. The raster of land cover.}

\item{dsm}{Raster. The digital surface model(DSM) that is used for
function 'calculate_viewshed' to calculate viewshed.}

\item{visiblepoints}{Dataframe. The viewshed calulated by
function 'calculate_viewshed'.}

\item{vegetation}{Numeric. The code of perviousness including trees or grass
in the raster of land cover.}

\item{imperviousness}{Numeric. The code of imperviousness including
buildings, or parkings or paths/roads in the raster of land cover.}
}
\value{
Vector. Area (sq ft) of perviousness and imperviousness. Percentages
of perviousness and imperviousness.
}
\description{
calculate_landcover
}
