% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_viewshed.R
\name{calculate_viewshed}
\alias{calculate_viewshed}
\title{calculate_viewshed}
\usage{
calculate_viewshed(
  dsm,
  under = NULL,
  dem = NULL,
  viewpoint,
  sample_points = NULL,
  r = NULL
)
}
\arguments{
\item{dsm}{the raster layer of digital surface model/digital elevation model}

\item{under}{the raster layer of digital surface model where the elevation
above crown base of trees was removed}

\item{dem}{the raster of terrain without any vertical object. If input of
dsm is already digital elevation model, this argument can be ignored}

\item{viewpoint}{a matrix including x,y coordinates}

\item{sample_points}{a matrix of points that are converted from DSM.
This can be ignored. (it is defaulted as NULL)}

\item{r}{the radius for viewshed analysis. (it is defaulted as NULL)}
}
\value{
Dataframe. The output dataframe includes x anf y coordinates.
}
\description{
calculate_viewshed
}
\examples{

}
