% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\title{get_extent}
\usage{
get_extent(dsm, visiblepoints)
}
\arguments{
\item{dsm}{The digital surface model(DSM) that is used for
function 'calculate_viewshed' to calculate viewshed.}

\item{visiblepoints}{the viewshed calulated by function 'calculate_viewshed'.}
}
\value{
Numeric. The area of viewshed extent.
}
\description{
get_extent
}
