% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_horizontal.R
\name{get_horizontal}
\alias{get_horizontal}
\title{get_horizontal}
\usage{
get_horizontal(dsm, dem, visiblepoints, type)
}
\arguments{
\item{dsm}{The digital surface model(DSM) that is used for
function 'calculate_viewshed' to calculate viewshed.}

\item{dem}{The raster of terrain without any vertical object}

\item{visiblepoints}{the viewshed calulated by function 'calculate_viewshed'.}

\item{type}{The metric to be returned. 1=depth(Furthest distance);
2=depth variation; 3=both (a list).}
}
\value{
Numeric. If type = 1 or 2, the function will return area of
gound surface or standard deviation of elevations as a number. If type = 3,
the function will return both metrics above as a vestor.
}
\description{
get_horizontal
}
