% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lidar.R
\name{get_lidar}
\alias{get_lidar}
\title{get_lidar}
\usage{
get_lidar(x, y, r, epsg, folder, plot = FALSE)
}
\arguments{
\item{x}{numeric, indicating Longtitude of the center point.}

\item{y}{numeric, indicating latitude of the center point.}

\item{r}{numeric, indicating searching distance.
The maximum distance is 800m (2625ft).
If r > 800m (2625ft), it will be reset to 800m (2625ft).}

\item{epsg}{numeric, indicating epsg code}

\item{folder}{string, indicating a path for downloading the LiDAR data}

\item{plot}{logical, if it is TRUE, final data will be visualized as
point cloud. The default is FALSE}
}
\value{
.las data
}
\description{
Search for and download LiDAR data based on coordinates
of a spatial point with a given distance. The maximum distance is 800m.
Different dataset could be found and the function automatically
To get more details of data on a larger scale,
please use viewscape::lidar_search.
}
\references{
Jean-Romain Roussel and David Auty (2022).
Airborne LiDAR Data Manipulation and Visualization for
Forestry Applications. R package version 4.0.1. https://cran.r-project.org/package=lidR
}
