% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_search.R
\name{lidar_search}
\alias{lidar_search}
\title{lidar_search}
\usage{
lidar_search(bbox, preview = FALSE, folder = NULL)
}
\arguments{
\item{bbox}{vector, indicating the bounding box to search for data.}

\item{preview}{logical. If TRUE, the image of the LiDAR data will
be display for the preview purposes. The default is FALSE}

\item{folder}{string, indicating a path for downloading the LiDAR data.
If folder is set, All available data will be downloaded.}
}
\value{
dataframe
}
\description{
Search for data within a bounding box via
the TNMAccess API: https://tnmaccess.nationalmap.gov/api/v1/docs
}
