% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_mask.R
\name{sector_mask}
\alias{sector_mask}
\title{sector_mask}
\usage{
sector_mask(viewshed, fov)
}
\arguments{
\item{viewshed}{viewshed object, generated by compute_viewshed()}

\item{fov}{Vector, specifying the field of view with two angles in degree (e.g. c(10,100))
for masking a viewshed based on its viewpoints.
See details.}
}
\value{
viewshed object
}
\description{
The sector_mask function is designed to subset a viewshed based on
its viewpoint and the field of view
}
\details{
For defining the field of view ('fov'), angles range from 0 to 360 degrees,
with 0 inclusive and 360 exclusive. The initial angle must be smaller than the terminal
angle in the sequence c(a,b) (a < b). To capture the northeast quadrant of a viewshed,
one would use c(0,90), while the eastern quadrant would be delineated by c(45,315) as

shown below:
135  90  45
180   v   0
225 270 315
Here, 'v' represents the viewpoint, with angles measured clockwise from due north.
}
