% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Viewshed.R
\docType{class}
\name{Viewshed-class}
\alias{Viewshed-class}
\title{An S4 class to represent the viewshed}
\description{
A \code{viewshed} object contains a 'matrix' of visible and invisible area,
resolution, extent, and crs
}
\section{Slots}{

\describe{
\item{\code{visible}}{matrix}

\item{\code{resolution}}{vector}

\item{\code{extent}}{numeric}

\item{\code{crs}}{character}
}}

