% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_viewshed.R
\name{visualize_viewshed}
\alias{visualize_viewshed}
\title{visualize_viewshed}
\usage{
visualize_viewshed(
  viewshed,
  extent,
  dsm = NULL,
  plot = "polygon",
  output = FALSE,
  type = NULL
)
}
\arguments{
\item{viewshed}{the viewshed dataset of a single viewpoint from
function calculate viewshed}

\item{dsm}{the raster layer of digital surface model that has been used
to calcualte viewshed}

\item{plot}{the type of visualization of a viewshed,
including "polygon", "raster", and "3D".}

\item{output}{if TRUE, the viewshed visualization will be returned.}

\item{type}{the type of visualized viewshed format to be returned,
including "raster" and "polygon".}
}
\value{
Raster or Polygon of a viewshed
}
\description{
visualize_viewshed
}
\examples{

}
