% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_feature.R
\name{calculate_feature}
\alias{calculate_feature}
\title{calculate_feature}
\usage{
calculate_feature(type, feature, viewshed, exclude_value = 0)
}
\arguments{
\item{type}{Numeric. The input type of canopy raster.
type=1: percentage raster (that represents the percentage of
canopy area in each cell).
type=2: binary raster (that only uses two values to represent whether
canopy exists in each cell).}

\item{feature}{Raster.}

\item{viewshed}{Viewshed object.}

\item{exclude_value}{Numeric. If type = 2, it indicates the value of
cells, where the feature doesn't exist.}
}
\value{
Numeric. The canopy area in the viewshed.
}
\description{
The calculate_feature function is designed to extract specific
feature-related information within a viewshed. It allows you to compute
the proportion of the feature that is present in the viewshed.
}
