% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_viewshed.R
\name{compute_viewshed}
\alias{compute_viewshed}
\title{compute_viewshed}
\usage{
compute_viewshed(
  dsm,
  viewpoints,
  offset_viewpoint = 1.7,
  offset_height = 0,
  r = NULL,
  multiviewpoints = FALSE,
  parallel = FALSE,
  raster = FALSE,
  plot = FALSE
)
}
\arguments{
\item{dsm}{raster, the digital surface model/digital elevation model}

\item{viewpoints}{vector, including x,y coordinates of a viewpoint
or a matrix including several viewpoints with x,y coordinates (if multiviewpoints = TRUE)}

\item{offset_viewpoint}{numeric, setting the height of the viewpoint.}

\item{offset_height}{numeric, setting the height of positions that a given viewpoint will
look at. The defaut is 0}

\item{r}{numeric, setting the radius for viewshed analysis. (it is defaulted as NULL)}

\item{multiviewpoints}{the radius for viewshed analysis. (it is defaulted as NULL)}

\item{parallel}{logical, indicating if parallel computing should be used to compute
viewsheds of multiview points. The default is FALSE. When it is TRUE, arguements 'raster'
and 'plot' are ignored}

\item{raster}{logical, if it is TRUE, the raster of viewshed will be returned.
The default is FALSE}

\item{plot}{logical, if it is TRUE, the raster of viewshed will be displayed}
}
\value{
Raster or list. If raster is TRUE, the output is a binary raster.
Value 1 means visible while value 0 means invisible. The list includes a binary matrix,
where Value 1 means visible while value 0 means invisible, and the extent of the vewshed.
If parallel is TRUE, the output is the list and visualization is unavailable.
}
\description{
Compute the viewshed based on a given spaial point or a set of points.
}
\details{
Parallel computing used the functions from BiocParallel package
}
\examples{
test_viewpoint <- c(test_viewpoint[,1], test_viewpoint[,2])
#Compute viewshed
output <- compute_viewshed(dsm = dsm,
                           viewpoints = test_viewpoint,
                           offset_viewpoint = 6,
                           raster = TRUE,
                           plot = TRUE)
}
\references{
Martin Morgan, Jiefei Wang, Valerie Obenchain, Michel Lang,
Ryan Thompson and Nitesh Turaga (2021). BiocParallel: Bioconductor facilities
for parallel evaluation. R package version 1.28.3.
https://github.com/Bioconductor/BiocParallel
}
