% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{get_depth}
\usage{
get_depth(visiblepoints, viewpoint, type)
}
\arguments{
\item{visiblepoints}{The viewshed calulated by function 'calculate_viewshed'}

\item{viewpoint}{A matrix including x,y coordinates}

\item{type}{The metric to be returned. 1=depth(Furthest distance);
2=depth variation; 3=both (a list).}
}
\value{
Numeric. If type = 1 or 2, the function will return furthest distance
or standard deviation of distance as a number. If type = 3, the function will
return both metrics above as a vector.
}
\description{
get_depth
}
