% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lidar.R
\name{get_lidar}
\alias{get_lidar}
\title{get_lidar}
\usage{
get_lidar(x, y, r, epsg, max_return, folder, plot = FALSE)
}
\arguments{
\item{x}{numeric, indicating Longtitude of the center point.}

\item{y}{numeric, indicating latitude of the center point.}

\item{r}{numeric, indicating search distance for LiDAR data.
The maximum distance is 1000m (3281ft).
If r > 1000m, it will be reset to 1000m.}

\item{epsg}{numeric, the EPSG code specifying the coordinate reference system.}

\item{max_return}{numeric, indicating the maximum of returns.}

\item{folder}{string, indicating a path for downloading the LiDAR data}

\item{plot}{logical (default is FALSE), enable or disable the plotting of
the processed LiDAR data.}
}
\value{
The function returns the processed LiDAR data as a lidR LAS object.
}
\description{
Search for and download LiDAR data based on coordinates
of a spatial point with a given distance. The maximum distance is 800m.
Different dataset could be found and the function automatically downloads
the latest dataset
To get more details of data on a larger scale, please use viewscape::lidar_search.
}
\references{
Jean-Romain Roussel and David Auty (2022).
Airborne LiDAR Data Manipulation and Visualization for
Forestry Applications. R package version 4.0.1. https://cran.r-project.org/package=lidR
}
