% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pano_view.R
\name{pano_view}
\alias{pano_view}
\title{pano_view}
\usage{
pano_view(
  dsm = NULL,
  semantic = NULL,
  vpt = NULL,
  h = 6,
  method = "equirectangular",
  sky_threshold = 3,
  step_size = 0.5,
  max_dist = 500,
  pano_dim = c(256, 512),
  heading = 0,
  plot = FALSE,
  legend = TRUE,
  axes = TRUE
)
}
\arguments{
\item{dsm}{A SpatRaster object (single layer) of surface elevation.}

\item{semantic}{A SpatRaster object (single layer) of semantic values such as
land cover and land use.}

\item{vpt}{A POINT sf object representing the viewpoint
(in projected coordinates).}

\item{h}{Numeric. Height of the viewpoint above the ground surface (in meters).}

\item{method}{character. \code{cylindrical} or \code{equirectangular}}

\item{sky_threshold}{Numeric. Elevation buffer (in meters) to consider a ray
reaching sky (default: 3.0).}

\item{step_size}{Numeric. Distance (in pixels) between sampling steps
(default: 0.5).}

\item{max_dist}{Numeric. Maximum ray distance (default: 500).}

\item{pano_dim}{Integer vector of size 2: height and width of the output
panorama (default: c(256, 512)).}

\item{heading}{numeric. Horizontal direction in degree. (default: 0). See Details.}

\item{plot}{logical. Whether to plot result. (default: FLASE)}

\item{legend}{logical. Whether to display lengend when plotting the result.
(default: TRUE)}

\item{axes}{logical. Whether to display axis when plotting the result.
(default: TRUE)}
}
\value{
A SpatRaster object representing the distance panorama, with \code{NA}
where sky is detected.
}
\description{
Generate tangential panoramic view from a DSM based on the Geo-transformation
method proposed by Lu et al. (2020).
}
\details{
For \code{heading}:
\itemize{
\item heading = 0 → facing north (default)
\item heading = 90 → facing east
\item heading = 180 → facing south
\item heading = 270 → facing due west
}
}
\examples{
dsm <- terra::rast(system.file("test_dsm.tif", package = "viewscape"))
vpt <- sf::read_sf(system.file("test_viewpoint.shp", package = "viewscape"))
result <- streetscape::pano_view(dsm, vpt, 6, method = 'cylindrical')

}
\references{
Lu, X., Li, Z., Cui, Z., Oswald, M. R., Pollefeys, M., & Qin, R. (2020).
Geometry-aware satellite-to-ground image synthesis for urban areas.
In Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern
Recognition (pp. 859-867).
}
