% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_magnitude.R
\name{visual_magnitude}
\alias{visual_magnitude}
\title{visual_magnitude}
\usage{
visual_magnitude(dsm, viewshed)
}
\arguments{
\item{dsm}{Raster, the digital surface / elevation model}

\item{viewshed}{Viewshed object.}
}
\value{
raster
}
\description{
Visual Magnitude quantifies the extent of a visible region
as perceived by an observer. It is derived from the surface's slope and
angle features, alongside the observer's relative distance from the area
in question.
}
\references{
Chamberlain, B. C., & Meitner, M. J. (2013).
A route-based visibility analysis for landscape management.
Landscape and Urban Planning, 111, 13-24.
}
